split up functions into more concrete scripts
